initial setup (#2)
## Purpose

This PR adds Elixir and Postgres features to the dev container in the project. The devcontainer.json file has been updated to include the necessary images for these features. This change will allow the developer to develop and test their Elixir application with Postgres as the database.

## Approach

The devcontainer.json file has been updated to include the ghcr.io/devcontainers-contrib/features/elixir-asdf:2 and ghcr.io/devcontainers-contrib/features/postgres-asdf:1 images for Elixir and Postgres respectively. Additionally, the image has been updated to the latest version of the universal dev container (mcr.microsoft.com/devcontainers/universal:2). These changes will enable the developer to work with Elixir and Postgres in the dev container.

## Commits

- add credo, dyalixir and excoveralls
- add CI
- add module docs
- update erlang and elixir
- add devcontainer
- format files
- modify gitignore
- make credo happy